Move image to new dir
With the intention of good practice and
for better organisation, I created a new
directory for all images. For e.g, if I
wanted to develop a website like this it
would be likely that I would need more images.
Before this change, I had the image stored
in the same directory as the page that used
it. However, in the future, multiple pages
may use the same image. Therefore, I created
directory dedicated to images alone.