Implement style rule for ingredients list
I initially implemented a style rule which set the bullet points on the
home page to a large font. However, as a result of linking all pages to
the stylesheet I failed to realise that bullet points were used also on
the recipe pages. Thus I created a style rule using a class selector
overriding the element selector.